Adding a simple configuration system
The mod will look for a configration file in a fixed location, and will use that file to set certain parameters (like the location of the event log) if it is present. This is mostly to assist with some development concerns, e.g. being able to dump the event log to a custom location for ease of use.